requirements.txt
asn1crypto==0.24.0
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
crcmod==1.7
cryptography==36.0.0
entrypoints==0.3
future==0.18.2
google-api-core==2.3.2
google-auth==2.3.3
google-cloud-bigquery==2.31.0
google-cloud-bigquery-storage==2.10.1
google-cloud-core==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
greenlet==1.1.2
grpcio==1.43.0
grpcio-status==1.43.0
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
keyring==17.1.1
keyrings.alt==3.1.1
libcst==0.3.23
mypy-extensions==0.4.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
py==1.11.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
PyGObject==3.30.4
pyOpenSSL==21.0.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
pytz==2021.3
pyxdg==0.25
PyYAML==6.0
requests==2.26.0
rsa==4.8
SecretStorage==2.3.1
six==1.12.0
SQLAlchemy==1.4.29
sqlalchemy-bigquery==1.2.2
toml==0.10.2
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
zipp==3.6.0
